@@ -1,10 +1,12 @@
 package service
 
 import (
-	"io/ioutil"
+	"io"
+	"os"
 	"strings"
 
 	"github.com/coreos/go-systemd/v22/unit"
+	"github.com/golang/glog"
 	"gopkg.in/yaml.v2"
 )
 
@@ -28,7 +30,7 @@ OUTER:
 				continue OUTER
 			}
 		}
-
+		glog.Infof("DEBUG: %+v %v", optsA, *optB)
 		return true, nil
 	}
 
@@ -54,7 +56,7 @@ OUTER:
 		newServiceOptions = append(newServiceOptions, opt)
 	}
 
-	data, err := ioutil.ReadAll(unit.Serialize(newServiceOptions))
+	data, err := io.ReadAll(unit.Serialize(newServiceOptions))
 	if err != nil {
 		return nil, err
 	}
@@ -83,7 +85,7 @@ OUTER:
 		serviceOptions = append(serviceOptions, appendOpt)
 	}
 
-	data, err := ioutil.ReadAll(unit.Serialize(serviceOptions))
+	data, err := io.ReadAll(unit.Serialize(serviceOptions))
 	if err != nil {
 		return nil, err
 	}
@@ -97,7 +99,7 @@ OUTER:
 
 // ReadServiceInjectionManifestFile reads service injection file
 func ReadServiceInjectionManifestFile(path string) (*Service, error) {
-	data, err := ioutil.ReadFile(path)
+	data, err := os.ReadFile(path)
 	if err != nil {
 		return nil, err
 	}
@@ -116,7 +118,7 @@ func ReadServiceInjectionManifestFile(path string) (*Service, error) {
 
 // ReadServiceManifestFile reads service file
 func ReadServiceManifestFile(path string) (*Service, error) {
-	data, err := ioutil.ReadFile(path)
+	data, err := os.ReadFile(path)
 	if err != nil {
 		return nil, err
 	}
@@ -135,7 +137,7 @@ func ReadServiceManifestFile(path string) (*Service, error) {
 
 // ReadScriptManifestFile reads script file
 func ReadScriptManifestFile(path string) (*ScriptManifestFile, error) {
-	data, err := ioutil.ReadFile(path)
+	data, err := os.ReadFile(path)
 	if err != nil {
 		return nil, err
 	}
