@@ -94,7 +94,7 @@ func newNodePolicy() *v1.SriovNetworkNodePolicy {
 	}
 }
 
-func newVdpaNodePolicy() *v1.SriovNetworkNodePolicy {
+func newVirtioVdpaNodePolicy() *v1.SriovNetworkNodePolicy {
 	return &v1.SriovNetworkNodePolicy{
 		ObjectMeta: metav1.ObjectMeta{
 			Name: "p1",
@@ -102,6 +102,29 @@ func newVdpaNodePolicy() *v1.SriovNetworkNodePolicy {
 		Spec: v1.SriovNetworkNodePolicySpec{
 			DeviceType: consts.DeviceTypeNetDevice,
 			VdpaType:   consts.VdpaTypeVirtio,
+			NicSelector: v1.SriovNetworkNicSelector{
+				PfNames:     []string{"ens803f1#2-3"},
+				RootDevices: []string{"0000:86:00.1"},
+				Vendor:      "8086",
+			},
+			NodeSelector: map[string]string{
+				"feature.node.kubernetes.io/network-sriov.capable": "true",
+			},
+			NumVfs:       4,
+			Priority:     99,
+			ResourceName: "virtiovdpa",
+		},
+	}
+}
+
+func newVhostVdpaNodePolicy() *v1.SriovNetworkNodePolicy {
+	return &v1.SriovNetworkNodePolicy{
+		ObjectMeta: metav1.ObjectMeta{
+			Name: "p1",
+		},
+		Spec: v1.SriovNetworkNodePolicySpec{
+			DeviceType: consts.DeviceTypeNetDevice,
+			VdpaType:   consts.VdpaTypeVhost,
 			NicSelector: v1.SriovNetworkNicSelector{
 				PfNames:     []string{"ens803f1"},
 				RootDevices: []string{"0000:86:00.1"},
@@ -112,7 +135,7 @@ func newVdpaNodePolicy() *v1.SriovNetworkNodePolicy {
 			},
 			NumVfs:       2,
 			Priority:     99,
-			ResourceName: "p1res",
+			ResourceName: "vhostvdpa",
 		},
 	}
 }
@@ -438,6 +461,55 @@ func TestSriovNetworkNodePolicyApply(t *testing.T) {
 			},
 		},
 		{
+			// vdpa policy with same priority (both virtio and vhost), VfRange's do not overlap so all is merged
+			tname: "one vdpa policy present same pf same priority partitioning",
+			currentState: func() *v1.SriovNetworkNodeState {
+				st := newNodeState()
+				st.Spec.Interfaces = []v1.Interface{
+					{
+						Name:       "ens803f1",
+						NumVfs:     4,
+						PciAddress: "0000:86:00.1",
+						VfGroups: []v1.VfGroup{
+							{
+								DeviceType:   consts.DeviceTypeNetDevice,
+								VdpaType:     consts.VdpaTypeVhost,
+								ResourceName: "vhostvdpa",
+								VfRange:      "0-1",
+								PolicyName:   "p2",
+							},
+						},
+					},
+				}
+				return st
+			}(),
+			policy: newVirtioVdpaNodePolicy(),
+			equalP: true,
+			expectedInterfaces: []v1.Interface{
+				{
+					Name:       "ens803f1",
+					NumVfs:     4,
+					PciAddress: "0000:86:00.1",
+					VfGroups: []v1.VfGroup{
+						{
+							DeviceType:   consts.DeviceTypeNetDevice,
+							VdpaType:     consts.VdpaTypeVirtio,
+							ResourceName: "virtiovdpa",
+							VfRange:      "2-3",
+							PolicyName:   "p1",
+						},
+						{
+							DeviceType:   consts.DeviceTypeNetDevice,
+							VdpaType:     consts.VdpaTypeVhost,
+							ResourceName: "vhostvdpa",
+							VfRange:      "0-1",
+							PolicyName:   "p2",
+						},
+					},
+				},
+			},
+		},
+		{
 			// policy with same priority that overwrites the 2 present groups in
 			// SriovNetworkNodeState because they overlap VfRange
 			tname: "two policy present same pf same priority overlap",
@@ -637,7 +709,7 @@ func TestSriovNetworkNodePolicyApply(t *testing.T) {
 	}
 }
 
-func TestVdpaNodePolicyApply(t *testing.T) {
+func TestVirtioVdpaNodePolicyApply(t *testing.T) {
 	testtable := []struct {
 		tname              string
 		currentState       *v1.SriovNetworkNodeState
@@ -647,9 +719,56 @@ func TestVdpaNodePolicyApply(t *testing.T) {
 		expectedErr        bool
 	}{
 		{
-			tname:        "vdpa configuration",
+			tname:        "virtio/vdpa configuration",
 			currentState: newNodeState(),
-			policy:       newVdpaNodePolicy(),
+			policy:       newVirtioVdpaNodePolicy(),
+			equalP:       false,
+			expectedInterfaces: []v1.Interface{
+				{
+					Name:       "ens803f1",
+					NumVfs:     4,
+					PciAddress: "0000:86:00.1",
+					VfGroups: []v1.VfGroup{
+						{
+							DeviceType:   consts.DeviceTypeNetDevice,
+							VdpaType:     consts.VdpaTypeVirtio,
+							ResourceName: "virtiovdpa",
+							VfRange:      "2-3",
+							PolicyName:   "p1",
+						},
+					},
+				},
+			},
+		},
+	}
+	for _, tc := range testtable {
+		t.Run(tc.tname, func(t *testing.T) {
+			err := tc.policy.Apply(tc.currentState, tc.equalP)
+			if tc.expectedErr && err == nil {
+				t.Errorf("Apply expecting error.")
+			} else if !tc.expectedErr && err != nil {
+				t.Errorf("Apply error:\n%s", err)
+			}
+			if diff := cmp.Diff(tc.expectedInterfaces, tc.currentState.Spec.Interfaces); diff != "" {
+				t.Errorf("SriovNetworkNodeState spec diff (-want +got):\n%s", diff)
+			}
+		})
+	}
+}
+
+func TestVhostVdpaNodePolicyApply(t *testing.T) {
+	testtable := []struct {
+		tname              string
+		currentState       *v1.SriovNetworkNodeState
+		policy             *v1.SriovNetworkNodePolicy
+		expectedInterfaces v1.Interfaces
+		equalP             bool
+		expectedErr        bool
+	}{
+		{
+			tname:        "vhost/vdpa configuration",
+			currentState: newNodeState(),
+			policy:       newVhostVdpaNodePolicy(),
 			equalP:       false,
 			expectedInterfaces: []v1.Interface{
 				{
@@ -659,8 +778,8 @@ func TestVdpaNodePolicyApply(t *testing.T) {
 					VfGroups: []v1.VfGroup{
 						{
 							DeviceType:   consts.DeviceTypeNetDevice,
-							VdpaType:     consts.VdpaTypeVirtio,
-							ResourceName: "p1res",
+							VdpaType:     consts.VdpaTypeVhost,
+							ResourceName: "vhostvdpa",
 							VfRange:      "0-1",
 							PolicyName:   "p1",
 						},
