@@ -2,80 +2,59 @@ package conformance
 
 import (
 	"flag"
-	"fmt"
-	"os"
+	"log"
 	"path"
 	"testing"
-
-	"github.com/k8snetworkplumbingwg/sriov-network-operator/test/util/clean"
+	"time"
 
 	. "github.com/onsi/ginkgo/v2"
-	"github.com/onsi/ginkgo/v2/reporters"
-	"github.com/onsi/ginkgo/v2/types"
 	. "github.com/onsi/gomega"
 
-	testclient "github.com/k8snetworkplumbingwg/sriov-network-operator/test/util/client"
+	kniK8sReporter "github.com/openshift-kni/k8sreporter"
 
 	// Test files in this package must not end with `_test.go` suffix, as they are imported as go package
 	_ "github.com/k8snetworkplumbingwg/sriov-network-operator/test/conformance/tests"
 
+	"github.com/k8snetworkplumbingwg/sriov-network-operator/test/util/clean"
 	"github.com/k8snetworkplumbingwg/sriov-network-operator/test/util/k8sreporter"
 )
 
 var (
-	junitPath      *string
-	dumpOutput     *bool
-	reporterFile   string
-	customReporter *k8sreporter.KubernetesReporter
+	customReporter *kniK8sReporter.KubernetesReporter
+	err            error
+	reportPath     *string
 )
 
 func init() {
-	dumpOutput = flag.Bool("dump", false, "dump informations for failed tests")
-	junitPath = flag.String("junit", "", "the path for the junit format report")
+	reportPath = flag.String("report", "", "the path of the report directory containing details for failed tests")
 }
 
 func TestTest(t *testing.T) {
-	RegisterFailHandler(Fail)
+	// We want to collect logs before any resource is deleted in AfterEach, so we register the global fail handler
+	// in a way such that the reporter's Dump is always called before the default Fail.
+	RegisterFailHandler(
+		func(message string, callerSkip ...int) {
+			if customReporter != nil {
+				customReporter.Dump(10*time.Minute, CurrentSpecReport().FullText())
+			}
 
-	reporterFile = os.Getenv("REPORTER_OUTPUT")
+			// Ensure failing line location is not affected by this wrapper
+			for i := range callerSkip {
+				callerSkip[i]++
+			}
+			Fail(message, callerSkip...)
+		})
 
-	clients := testclient.New("")
-
-	if reporterFile != "" {
-		f, err := os.OpenFile(reporterFile, os.O_APPEND|os.O_CREATE|os.O_WRONLY, 0644)
+	if *reportPath != "" {
+		reportFile := path.Join(*reportPath, "sriov_failure_report.log")
+		customReporter, err = k8sreporter.New(reportFile)
 		if err != nil {
-			fmt.Fprintf(os.Stderr, "failed to open the file: %v\n", err)
-			return
+			log.Fatalf("Failed to create the k8s reporter %s", err)
 		}
-		defer f.Close()
-		customReporter = k8sreporter.New(clients, f)
-	} else if *dumpOutput {
-		customReporter = k8sreporter.New(clients, os.Stdout)
 	}
 
 	RunSpecs(t, "SRIOV Operator conformance tests")
 }
-
-var _ = ReportAfterSuite("conformance", func(report types.Report) {
-	if *junitPath != "" {
-		junitFile := path.Join(*junitPath, "junit_sriov_conformance.xml")
-		reporters.GenerateJUnitReportWithConfig(report, junitFile, reporters.JunitReportConfig{
-			OmitTimelinesForSpecState: types.SpecStatePassed | types.SpecStateSkipped,
-			OmitLeafNodeType:          true,
-			OmitSuiteSetupNodes:       true,
-		})
-	}
-})
-
-var _ = ReportAfterEach(func(sr types.SpecReport) {
-	if sr.Failed() == false {
-		return
-	}
-
-	if reporterFile != "" || *dumpOutput {
-		customReporter.Report(sr)
-	}
-})
 
 var _ = BeforeSuite(func() {
 	err := clean.All()
